+++ core/src/main/java/feign/Util.java
@@ -24,6 +24,8 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.Reader;
 import java.lang.reflect.Array;
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.lang.reflect.WildcardType;
@@ -103,6 +105,20 @@ public class Util {
     if (!expression) {
       throw new IllegalStateException(format(errorMessageTemplate, errorMessageArgs));
     }
+  }
+
+  /** Identifies a method as a default instance method. */
+  public static boolean isDefault(Method method) {
+    // Default methods are public non-abstract, non-synthetic, and non-static instance methods
+    // declared in an interface.
+    // method.isDefault() is not sufficient for our usage as it does not check
+    // for synthetic methods.  As a result, it picks up overridden methods as well as actual default
+    // methods.
+    final int SYNTHETIC = 0x00001000;
+    return ((method.getModifiers()
+                & (Modifier.ABSTRACT | Modifier.PUBLIC | Modifier.STATIC | SYNTHETIC))
+            == Modifier.PUBLIC)
+        && method.getDeclaringClass().isInterface();
   }
 
   /** Adapted from {@code com.google.common.base.Strings#emptyToNull}. */